requirement.txt
appdirs==1.4.2
coreapi==2.3.0
coreschema==0.0.4
coverage==4.0.3
defusedxml==0.5.0
Django==1.10.6
django-filter==1.0.1
django-guardian==1.4.6
django-jinja==2.2.2
django-rest-swagger==2.1.2
djangorestframework==3.5.4
djangorestframework-jwt==1.9.0
djet==0.2.1
djoser==0.5.4
itypes==1.1.0
Jinja2==2.9.5
MarkupSafe==0.23
oauthlib==2.0.1
openapi-codec==1.3.1
packaging==16.8
PyJWT==1.4.2
pyparsing==2.2.0
python-coveralls==2.9.0
python3-openid==3.1.0
PyYAML==3.12
requests==2.13.0
requests-oauthlib==0.8.0
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
